requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.129.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
imbalanced-learn==0.14.1
iniconfig==2.3.0
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.46.0b1
matplotlib==3.10.8
numba==0.63.0b1
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
